requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.1
django-allauth==0.39.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-sendgrid-v5==0.8.0
future==0.17.1
idna==2.8
oauthlib==3.0.1
Pillow==6.0.0
python-http-client==3.1.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
sendgrid==6.0.5
sqlparse==0.3.0
urllib3==1.25.3
